requirements.txt
langchain==0.0.352
openai==1.3.5
google-search-results==2.4.2
python-dotenv==1.0.0
langchain-experimental==0.0.47
unstructured==0.11.6
chromadb==0.4.21
tiktoken==0.5.2
langchainhub==0.1.14
bs4==0.0.1
pypdf==3.17.4


